Fix Argument<T>.GetMemberName
Method will now "properly" check that the target is indeed a compiler
generated class. This is at least a step in the right direction but it
still feels rather naive and very hackish.

As long as we're unable to get the actual IL or expression tree from the
function this will keep being hackish, maybe the method should just be
taken out instead.